refactor core AutoModel
