+++ gradsflow/callbacks/logger/comet.py
@@ -12,6 +12,8 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import os
+
+os.environ["COMET_DISABLE_AUTO_LOGGING"] = "1"
 from typing import TYPE_CHECKING, Optional
 
 BaseExperiment = None
