+++ gradsflow/data/autodata.py
@@ -11,7 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from typing import TYPE_CHECKING, Callable, Optional
+from typing import Callable, Optional
 
 from accelerate import Accelerator
 from loguru import logger
