+++ gradsflow/tasks/engine/__init__.py
@@ -11,16 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-import os
-import sys
-import warnings
-
-try:
-    import gradsflow
-except ImportError:
-    sys.path.append("./")
-
-os.environ["GF_CI"] = "true"
-
-warnings.filterwarnings("ignore")
