+++ tests/tasks/test_autotrainer.py
@@ -15,12 +15,12 @@ from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 
-from gradsflow.core.backend import AutoBackend
+from gradsflow.tasks.engine.backend import AutoBackend
 
 trainer_config = {"show_progress": False}
 
 
-@patch("gradsflow.core.backend.pl")
+@patch("gradsflow.tasks.engine.backend.pl")
 def test_optimization_objective(mock_pl: Mock):
     dm = MagicMock()
     model_builder = MagicMock()
